block: fix logic to determine DM partition kname on Ubuntu 25.10+
The logic we had to determine a partition kname for a DM device is based
on adding -part{n} to the first devlink of the device.

Before Ubuntu 25.10, the first devlink was typically
/dev/disk/by-id/dm-name-{name}, and the logic worked because
/dev/disk/by-id/dm-name-{name}-part{n} would exist

Unfortunately, since Ubuntu 25.10, the first devlink is
/dev/disk/by-diskseq/{seq} where seq is a number. The logic does not
work anymore since /dev/disk/by-diskseq/{seq}-part{n} is not a thing.

Fixed by looking specifically for the devlink which starts with
/dev/disk/by-id/dm-name-

LP: #2119429

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>